REFACTOR:  Minor refactoring of the RFFE Analyzer:
Added some comments and my trademark horizontal bars.
Adjusted the 'count' variable to reflect what it is, teh RFFE byte count.   Changed its type to U8 from S32 and modified appropriate methods.
Removed the clock_check method from the code for now as it is flawed in its logic and not a great test to begin with.   We might put in better thought out physical layer checks later.

Fiddled with the way the on-screen markers are drawn and what is drawn.  The only change of substance is that a down arrow is placed on the SCLK to indicate that these are the SDATA sample edges.