ENH:  Upgraded to handle unexpected SSCs.  Now an unexpected SSC halts the current packet processing and starts over.  This prevents the parsing from omiting valid packets that occur after a packet that is incomplete for any reason.
REFACTOR:  Adjusted which variables are member variables vs. local variables.   This went along with the above enhancement which modified the flow through the parsing slighlty.  The main change was that the current set of sample points were member variables but did not following the naming convention and were still passed around by some of the member variables.   The current set of result markers was passed around a lot as well, and since the sample points are store in member variable (scratchpad arrays really) I decided to do likewise for the results marker array.