BUG:  Fixed Data Value displayed in 0 byte writes (been wrong since the beginning as best I can tell) BUG:  Fixed a problem with the unexpected SSC Exception from recent checkins.  It was not caught when an SSC happens while you are halfway through finding an SSC.  Now interrupting an SSC with an SSC is also allowed and does not throw any uncaught exceptions. BUG:  Correctly printout all 16 bits of the Address in ExtLong cycles when exporting to a .CSV.  This was caught by inspection, but I have no way of testing this as yet.  (The generate traffic for simulation needs to be expanded to include all packet types and ideally introduce some error cases as well).
REFACTOR:  Never liked the way that FillInFrame worked.  Simplified its arguments by making use of all the member variables that were already there to begin with and a few more that I recently added.  (There was no need to pass in so many member variables to the method).   This streamlined the code quite a bit.   Since I was changing arguments I could also simplify some of the loops and such inside the function as well.
REFACTOR:  Fixed the way that 5 Bit, 8 Bit and 16 Bit addresses were categorized in the RFFEAnalyerResults typedefs.  Simplified the RFFEAnalyzerResults case statement as a result and displayed all the addresses as bytes (upper and lower in the ExtLong case) which is how we would think of them.